add an emission factors import option on seed script
